Adjust velocities and probability values of actors, improve enemy spawn logic
- Adjust velocities of all actors to scale dynmically according to target FPS
- Adjust probability values (spawn, fire, shield, etc) of all actors to scale dynmically according to target FPS
- Improved enemy spawn logic to allow better randomization of Y-axis placement

+++ src/main/java/com/example/demo/actors/Projectile.java
@@ -4,9 +4,9 @@ import com.example.demo.utils.LoggerUtil;
 
 public abstract class Projectile extends ActiveActorDestructible {
 
-    private final int horizontalVelocity;
+    private final double horizontalVelocity;
 
-	public Projectile(String imageName, int imageHeight, double initialXPos, double initialYPos, int horizontalVelocity) {
+	public Projectile(String imageName, int imageHeight, double initialXPos, double initialYPos, double horizontalVelocity) {
 		super(imageName, imageHeight, initialXPos, initialYPos);
 
         this.horizontalVelocity = horizontalVelocity;
